add http.MaxBytesReader to /sign
